requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.1.1
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pipx==1.4.3
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
userpath==1.9.1
uvicorn==0.29.0
